:sparkles: 메인 / 투자실험 소개 Page 스타일 반영
