+++ public/index.html
@@ -29,7 +29,7 @@
       name="google-signin-client_id"
       content="342833027510-ailtjkqe7b2vl34vl4c6tcpmm15chd24.apps.googleusercontent.com"
     />
-    <title>React App</title>
+    <title>finble</title>
   </head>
   <body>
     <noscript>You need to enable JavaScript to run this app.</noscript>
