+++ src/components/Button.tsx
@@ -67,7 +67,7 @@ const Btn60Container = styled.div<btn>`
       : 'var(--type-white)'};
   border: ${(props) =>
     props.type === 'login' || props.type === 'outline_disable'
-      ? '2px solid var(--type-gray-4)'
+      ? '1px solid var(--type-gray-4)'
       : props.type === 'outline_able'
       ? '2px solid var(--main-blue)'
       : 'none'};
